got votingContext file to work and load properly
