+++ my-app/src/pages/voting.js
@@ -1,13 +1,13 @@
-import React, { useState, useEffect, use } from 'react';
+import React, { useState, useEffect } from 'react';
 import {Text, Box, useDisclosure, Flex, Grid, Container, Spacer, VStack, Heading, Tabs, TabList, Tab, TabPanels, TabPanel, Button, Collapse, FormControl, FormLabel, Input, Textarea, RadioGroup, Stack, Radio, Modal, ModalOverlay, ModalContent, ModalHeader, ModalCloseButton, ModalBody, ModalFooter } from '@chakra-ui/react';
 
 
 import { ethers } from 'ethers';
-import KubixVotingABI from '../abi/KubixVoting.json';
-import KubidVotingABI from '../abi/KubidVoting.json';
+
 import { useDataBaseContext } from '@/contexts/DataBaseContext';
 import { useWeb3Context } from '@/contexts/Web3Context';
-import { useToast } from '@chakra-ui/react';
+import { useVoting } from '@/contexts/votingContext';
+
 import { BarChart, Bar, XAxis, YAxis} from 'recharts';
 import CountDown from '@/components/voting/countDown';
 
@@ -28,180 +28,28 @@ const glassLayerStyle = {
 
 
 const Voting = () => {
-  const {signerUniversal, providerUniversal, account}= useWeb3Context()
-  const { findMinMaxKubixBalance } = useDataBaseContext();
-
-  const contractX = new ethers.Contract('0x4Af0e1994c8e03414ffd523aAc645049bcdadbD6', KubixVotingABI.abi, signerUniversal);
-  const contractD = new ethers.Contract('0xaf395fbBdc0E2e99ae18D42F2724481BF1Ab02c8', KubidVotingABI.abi, signerUniversal);
 
 
-  const [proposal, setProposal] = useState({ name: '', description: '', execution: '', time: 0, options: [] ,id:0 });
+  const {setContract,contractX, contractD, contract, loadingVote, setLoadingVote, selectedPoll, setSelectedPoll,selectedOption, setSelectedOption, ongoingPollsKubix, setOngoingPollsKubix, completedPollsKubix, setCompletedPollsKubix, ongoingPollsKubid, setOngoingPollsKubid, completedPollsKubid, setCompletedPollsKubid, completedEnd, setCompletedEnd, totalCompletedCount, setTotalCompletedCount, proposal, setProposal, showCreateVote, setShowCreateVote, blockTimestamp, setBlockTimestamp, loadingSubmit, setLoadingSubmit, handleVote, createPoll, fetchPolls, fetchPollsData, loadMoreCompleted, handleSubmit, showCreatePoll, setShowCreatePoll } = useVoting();
+
+
+
+
   const [selectedTab, setSelectedTab] = useState(0);
-  const [showCreateVote, setShowCreateVote] = useState(false);
-  const [showCreatePoll, setShowCreatePoll] = useState(false);
-  const [blockTimestamp, setBlockTimestamp] = useState(0);
+
 
 
   const { isOpen, onOpen, onClose } = useDisclosure();
-  const [selectedPoll, setSelectedPoll] = useState(null);
-  const [selectedOption, setSelectedOption] = useState(null);
-
-
-  const [loadingSubmit, setLoadingSubmit] = useState(false)
-
-  const [loadingVote, setLoadingVote] = useState(false)
-
-  const [contract, setContract] = useState(contractX);
-
-  const [ongoingPollsKubix, setOngoingPollsKubix] = useState([]);
-  const [completedPollsKubix, setCompletedPollsKubix] = useState([]);
-  const [ongoingPollsKubid, setOngoingPollsKubid] = useState([]);
-  const [completedPollsKubid, setCompletedPollsKubid] = useState([]);
-
-
-  const [completedEnd, setCompletedEnd] = useState(3);
-
-  const [totalCompletedCount, setTotalCompletedCount] = useState(0);
 
 
 
 
 
-
-  const toast = useToast();
 
   const handlePollClick = (poll) => {
     setSelectedPoll(poll);
     onOpen();
   };
-
-  const handleVote = async () => {
-    setLoadingVote(true)
-    if (selectedOption === null) {
-      toast({
-        title: 'Error',
-        description: 'Please select an option to vote',
-        status: 'error',
-        duration: 3000,
-        isClosable: true,
-      });
-
-      setLoadingVote(false)
-      return;
-    }
-    
-    try {
-      // Call the vote function from the contract
-      console.log(selectedPoll.id, account, selectedOption[0])
-      console.log(account)
-      const tx = await contract.vote(selectedPoll.id, account, selectedOption[0]);
-      await tx.wait();
-      toast({
-        title: 'Vote submitted',
-        description: 'Your vote has been submitted successfully',
-        status: 'success',
-        duration: 3000,
-        isClosable: true,
-      });
-      onClose();
-    } catch (error) {
-      console.error(error);
-      toast({
-        title: 'Error',
-        description: 'There was an error submitting your vote. Please try again.',
-        status: 'error',
-        duration: 3000,
-        isClosable: true,
-      });
-    }
-
-    setLoadingVote(false)
-  };
-
-  const fetchPolls = async (selectedContract, setOngoingPollsFunc, setCompletedPollsFunc, completedStart = 0, completedEnd = 3) => {
-    try {
-        await selectedContract.moveToCompleted();
-
-        // Fetch both active and completed poll counts in parallel
-        const [ongoingPollsCount, completedPollsCount] = await Promise.all([
-            selectedContract.activeProposalsCount(),
-            selectedContract.completedProposalsCount() // Replace with your actual function if different
-        ]);
-
-        // Calculate how many polls to fetch
-        const completedEndModified = Math.min(completedEnd, completedPollsCount);
-
-        // Fetch ongoing and completed polls
-        const [ongoingPolls, completedPolls] = await Promise.all([
-          fetchPollsData(selectedContract, 0, ongoingPollsCount, false),
-          fetchPollsData(selectedContract, completedStart, completedEndModified, true)
-        ]);
-
-        setOngoingPollsFunc(ongoingPolls);
-        setCompletedPollsFunc(completedPolls);
-        
-    } catch (error) {
-        console.error(error);
-    }
-  };
-
-const fetchPollsData = async (selectedContract, start, end, completed) => {
-    const pollsPromises = Array.from({ length: end - start }, async (_, i) => {
-        
-        var index = start + i;
-        if (!completed){
-          index++
-        }
-
-        // Aligning the index to start from 0 for both completed and ongoing polls
-        const proposalId = await selectedContract[completed ? "completedProposalIndices" : "activeProposalIndices"](index);
-
-
-        const [proposal, optionsCount] = await Promise.all([
-            selectedContract.activeProposals(proposalId),
-            selectedContract.getOptionsCount(proposalId)
-        ]);
-
-        const pollOptionsPromises = Array.from({ length: optionsCount }, (_, j) => selectedContract.getOption(proposalId, j));
-
-        const pollOptions = await Promise.all(pollOptionsPromises);
-
-
-        let pollWithOptions = {
-            ...proposal,
-            options: pollOptions,
-            id: proposalId,
-            remainingTime: proposal.timeInMinutes * 60 - (Math.floor(Date.now() / 1000) - proposal.creationTimestamp)
-        };
-        
-        if (completed) {
-            const winner = await selectedContract.getWinner(index);
-            pollWithOptions = { ...pollWithOptions, winner };
-        }
-        
-        return pollWithOptions;
-    });
-
-    return await Promise.all(pollsPromises);
-};
-
-  const loadMoreCompleted = () => {
-    setCompletedEnd(prevCompletedEnd => prevCompletedEnd + 5);  // Increment completedEnd
-    try{
-    if (selectedTab === 0) {
-      fetchPolls(contractD, setOngoingPollsKubid, setCompletedPollsKubid, 0, completedEnd);
-
-    } else if (selectedTab === 1) {
-      fetchPolls(contractX, setOngoingPollsKubix, setCompletedPollsKubix, 0, completedEnd);
-
-    }
-  } catch (error) {
-    console.error(error);
-  }
-  };
-
-
-
 
 
 
@@ -211,37 +59,7 @@ const fetchPollsData = async (selectedContract, start, end, completed) => {
     setProposal({ ...proposal, [name]: value });
   };
 
-  const handleSubmit = async (e) => {
-    e.preventDefault();
-    setLoadingSubmit(true)
-    try {
-      console.log("handle submit")
-      await createPoll();
-      toast({
-        title: 'Poll created successfully',
-        description: 'Your poll has been created.',
-        status: 'success',
-        duration: 3000,
-        isClosable: true,
-      });
-      // Reset the form
-      setProposal({ name: '', description: '', execution: '', time: 0, options: [] });
-      setShowCreatePoll(false);
-      fetchPolls();
-//bugs: modal card dispalying last vote, glass modal overlay bad, ongoing votes doesnt have glass properly applied
-    } catch (error) {
-      console.error(error);
-      toast({
-        title: 'Error creating poll',
-        description: 'There was an error creating the poll. Please try again.',
-        status: 'error',
-        duration: 3000,
-        isClosable: true,
-      });
-    }
 
-    setLoadingSubmit(false)
-  };
 
   const handleTabsChange = (index) => {
     setSelectedTab(index);
@@ -260,31 +78,7 @@ const fetchPollsData = async (selectedContract, start, end, completed) => {
     asyncFunc()
   }, []);
 
-  const createPoll = async () => {
 
-    if (selectedTab == 1) {
-      
-    const balances = await findMinMaxKubixBalance();
-
-  
-    // Parse the options string into an array
-    const optionsArray = proposal.options.map(option => option.trim());
-  
-    const tx = await contract.createProposal(proposal.name, proposal.description, proposal.execution, balances.maxBalance, balances.minBalance, proposal.time, optionsArray);
-    await tx.wait();
-    } 
-    else {
-
-    
-      // Parse the options string into an array
-      const optionsArray = proposal.options.map(option => option.trim());
-    
-      const tx = await contract.createProposal(proposal.name, proposal.description, proposal.execution, proposal.time, optionsArray);
-      await tx.wait();
-
-    }
-
-  };
   
   const handleOptionsChange = (e) => {
     // Split the input string by comma to get an array of options
@@ -293,23 +87,13 @@ const fetchPollsData = async (selectedContract, start, end, completed) => {
   };
   
 
-  
-  async function fetchBlockTimestamp() {
-    const currentBlock = await providerUniversal.getBlock('latest');
-    setBlockTimestamp(currentBlock.timestamp);
-  }
 
-  useEffect(() => {
-    fetchBlockTimestamp();
-  }, []);
 
   const handleCreatePollClick = () => {
     setShowCreatePoll(!showCreatePoll);
   };
 
 
-
-  
   
 
   return (
